Shortened Client.js & Patched bugs
