+++ src/API/getOnline.js
@@ -1,0 +1,5 @@
+module.exports = async function () {
+  const res = await this._makeRequest('/playerCount');
+
+  return res.playerCount;
+};
